backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
asyncpg
alembic
pydantic
python-dotenv
websockets
langgraph
langchain-google-genai
langchain-core
langchain-community
python-multipart